[setup.py] use README.rst for 'long_description', used for PyPI project page
After publishing 1.0.6 via twine command, I noticed the PKG-INFO 'Description' was empty...
That's because setup.py was missing 'long_description' field.
So I had to manually upload a modified PKG-INFO with the content of the current README.rst.
For next releases this won't be necessary.

The 'License' was also missing in the sdist PKG-INFO, so I added that too.